Use pipelines in Ollama provider
We recently introduced Ollama as a provider but the pipelines were not
wired in.

We were querying Ollama directly in the endpoint without calling the
`complete` method of `BaseProvider`. This PR now calls `complete` from
Ollama provider. Potentially we would also need to add a `ModelOutputNormalizer`
as we did for LlamaCPP but that is left for a future PR.