Add an output pipeline
Adds an output pipeline with its own context and a possibility to buffer
in case a step tells the pipeline to not forward the chunk.

The buffer is stored into in case a chunk is not evaluated which means
that subsequent steps can keep on buffering potentially redacted text
for un-redaction later.

The remainder of the buffer is flushed upon consuming the stream.