Record the output of the LLM in the DB
We are going to store the complete output of the LLMs.
- If the response was stream then a list of JSON objects will get stored.
- If not stream then a single JSON object will get stored.

The PR also changes the initial schema. The schema will now also
store the complete input request to the LLM.

Storing the complete input and output should help debug any problem
with CodeGate as well as reproduce faithfully in the dashboard all
the conversations.