Change the output pipeline to return a list
The first version of the output pipeline returned either a ModelResponse
if the step was passing through a chunk or None if the chunk was to be
paused. For steps that enrich the streaming output, we want to inject
chunks into the output flow. Let's change the signature of the
processing to return a list of chunks, returning an empty list means
pause and if one or more chunks are returned, they are streamed to the
client.