Improve the code block regex and finding code snippets
Splits the regex into a multi-line one so that it is actually readable.
Changes the regex a bit so that the language is only the first word
until the first whitespace.

Handles code blocks that only contain the language which is common in
output snippets, e.g.:
```python
print("Hi")
```