Extract packages in both input and output using the LLM the user called with
Instead of always using the local LLM to extract the packages, use the
LLM the user called with.

This is implemented in a new module llm_utils - the way it is
implemented is a bit of a kludge as it really should reuse more of the
providers. But because of time constraints, let's start this way and
refactor later.

The important thing is the PackageExtractor class that uses the provided
string to extract packages out of the prompt.