Use `def` in FastAPI dashboard calls.
Quote from [FastAPI docs](https://fastapi.tiangolo.com/async/#path-operation-functions)
> Path operation functions
When you declare a path operation function with normal def instead of async def, it is run in an external threadpool that is then awaited, instead of being called directly (as it would block the server).

What we have experienced in Trusty is that if we use `async def` it will queue the requests and wait for them sequentially. While using just `def` will not block the server. This PR used `def` in the dashboard functions to not affect User Experience.