Adds cli command to generate certificate
This will add the code needed to generate certificates. It will
also add a command to our cli for generating certificates.

Our Dockerfile will also use the generate-certs command to be
executed before the server is deployed. Meaning we will always
have the certificates created in case they don't exist.

Note that for having the certificates at host we would need to
mount a volume to the container