start_serving should be False
According to the docs:

start_serving set to True (default) causes the created server to
start accepting connections immediately. When set to False,
the user should await Server.start_serving() or
Server.serve_forever() to make the server to start accepting
connections.

We call `await server.serve_forever()` in `run_proxy_server`
so we should allow this command to ditact we are able to accept
connections