Changed flow for recording objects on DB. Keep objects in `context`
With this change the objects that are going to be stored in DB are
kept in the `context` of the pipeline. The pipeline and its `context`
are used by all providers, including Copilot. We would need to find
a good place in Copilot provider to record the context in DB, e.g.
when all the chunks have been transmitted and the stream is about to
be closed.