Refine caching non-copilot FIM calls
Closes: #376

We were using the whole content of the prompt to hash the requests
that doesn't come from copilot. This is inefficient since the prompt
between requests can vary quite a lot. Instead use the filepath included
in every request. After some investigation, `copilot` puts the
filepath at the top of the prompt while the rest of the providers
include the filepath at the bottom of the context section.