Tune the system prompts to extract package names or ecosystems
With Copilot+Anthropic the user message can contain instructions that
are passed from the chat client and really sound like a system message.
For whatever reason, the chat client includes them with `role=user`.
These instructions confuse the LLM when extracting the package names or
ecosystem and it replies with a mix of JSON and natural language.

To work around that, Pankaj suggested to tune the system prompt and tell
the LLM to ignore the instructions. This seems to work and helps
Anthropic extract the package names even with copilot user message.

Co-authored-by: Pankaj Telang <pankaj@stacklok.com>

Related: #401