Add build-arg in `make image-build` command
The build-arg prevents that the step for downloading the UI code is cached.
This way we can ensure that we are always using the latest releasee from
the UI when building the image locally. The build-arg is already present
in the GH action to build and publish the image.

Additionally, remove the `secret` provided to the GH action to build the
image. It was used when the containers were not public