Add all-in-one minimal nginx config
The nginx configuration uses a lot of defaults that work well in
server-like environments. However, this is not optimized for a local
dev-environment or single-container as codegate is meant to run.

The nginx server spawns several processes which mostly remain idle. This
is wasteful and it is taking over resources that could be used for
something else. The intention of this new configuration is to minimize
footprint.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>