Set codegate version from build argument in container (#615)
The codegate container does not run the python program via poetry, thus
it does not have the necessary package metadata to determine the
version.

To work around this lack of information, this enhancement leverages
Docker build arguments to set a special variable that will set the
version. This way, we don't need to rely on package metadata nor
manually bump the version. This machinery was set up in both the local
image building and in CI, so we'll get the version and a commit hash.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>