Extended the CodeSnippetExtractor functionality for muxing (#945)
* Extended the CodeSnippetExtractor functionality for muxing

We used to have a generic CodeSnippetExtractor to detect code snippets
and the corresponding filepaths. This PR extends it's functionality
to be able to detect snippets coming from different clients.
After this PR the supported and tested clients are:
- Continue
- Cline
- Aider
- Open Interpreter

The main reason behind this change is to be able to catch the filepaths
coming through CodeGate, detect it's destination and mux according
to them. The PR also adds the necessary in the router used for muxing
to detect the filepaths from the request messages.

Since CodeSnippetExtractor is no longer a pipeline step it was moved
out of the folder `pipeline` to it's own folder

* Attended comments in PR, thanks Jakub