Fix for non streaming clients (#967)
* Fix for non-non streaming clients

Fixes bug that that returned improper json for non-streaming clients. This fix corrects also ensures that the chat requests and responses are persisted properly for display in codegate admin UI.

* Fix for non-streaming clients

Fixes  that returned improper json for non-streaming clients. This fix corrects also ensures that the chat requests and responses are persisted properly for display in codegate admin UI.