Record output from non-streaming responses in DB (#1078)
Closes: #938

We were not recording the output of the non-streamed responses
in the DB. In our API we're filtering out all the conversations that don't
contain an answer. Since we were not recording non-streamed responses
we were filtering out the entire conversation if it was non-streamed.

Note: This still doesn't add an output pipeline for non-streamed
responses.