fix: only record db content if it is the last chunk in stream (#1091)
For copilot chat, we are actually receiving multiple streams, and
we were recording entries and alerts for each one, repeating those.
So detect if we are in the last chunk and propagate it to the
pipeline, so we can record it only in this case, avoiding dupes.
In the case of other providers, we only receive one request, so
always force saving it

Closes: #936