Fix FIM for OpenRouter (#1097)
* Fix FIM for OpenRouter

FIM was not working with OpenRouter in Continue. The reason
was that we get FIM requests in `/completions`. LiteLLM when
using `acompletion` is forcing `/chat/completions` and the return
format `{..., "choices":[{"delta":{"content":"some text"}}]}`.
However, Continue was expecting the format:
`{..., "choices":[{"text":"some text"}]}` becuase of the endpoint
it called.

With this PR we force the return format to be the latter using
`atext_completion` from LiteLLM

* Force denormalization in OpenRouter FIM to have a prompt key