Add `instance` table along with init code. (#1234)
This change adds an `instance` table containing the minimum set of
details about codegate. We might want to add more details in the
future.

The table must contain only a single record at any time. To guarantee
that, a trigger is added that prevents inserts beyond the first
record.

Finally, code performing the initialization is added to the `serve`
command.